backend/requirements.txt
django
djangorestframework
django-cors-headers
django-filter
langchain-google-genai
langchain-community
langchain
langchain-openai
gunicorn
dj-database-url
psycopg2-binary
whitenoise
google-generativeai
reportlab
